refactor(project): unify numeric fields to long and inline part quota status

+++ src/main/java/com/umc/product/project/adapter/in/web/dto/common/ApplicationQuestionItem.java
@@ -9,7 +9,7 @@ import lombok.Builder;
  */
 @Builder
 public record ApplicationQuestionItem(
-    int order,
+    long order,
     QuestionType type,
     String title,
     boolean required,
